+++ tests/e2e/config/jest.setup.js
@@ -10,23 +10,6 @@ import {
 const { merchant } = require( '@woocommerce/e2e-utils' );
 
 /**
- * Add an expect range matcher
- * @see https://jestjs.io/docs/expect#expectextendmatchers
- */
-expect.extend({
-	toBeInRange: function (received, floor, ceiling) {
-		const pass = received >= floor && received <= ceiling;
-		const condition = pass ? 'not to be' : 'to be';
-
-		return {
-			message: () =>
-				`expected ${received} ${condition} within range ${floor} - ${ceiling}`,
-			pass,
-		};
-	},
-});
-
-/**
  * Navigates to the post listing screen and bulk-trashes any posts which exist.
  *
  * @return {Promise} Promise resolving once posts have been trashed.